+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+### Added
+- GFF3 parsing now supports providing a pre-built gffutils database
+
 ## [0.15.0] 2022-3-3
 ### Changed
 - Added flag `allow_duplicate_sequence_identifiers` to GenBank parser